iBooks related hacks....
iBooks does not allow the margin/padding modification of the `<body>` element. The `Overview.xhtml` file must be modified by adding a top level `<div>` element to encapsulate all content, and be used for padding...
