+++ docs/assets/js/search.js
@@ -1,3 +1,3 @@
 var typedoc = typedoc || {};
             typedoc.search = typedoc.search || {};
-            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"lib/fetch\"","url":"modules/_lib_fetch_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":4194304,"name":"URL","url":"modules/_lib_fetch_.html#url","classes":"tsd-kind-type-alias tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":2,"kind":32,"name":"text_content","url":"modules/_lib_fetch_.html#text_content","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":3,"kind":64,"name":"check_Web_url","url":"modules/_lib_fetch_.html#check_web_url","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":4,"kind":32,"name":"my_fetch","url":"modules/_lib_fetch_.html#my_fetch","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":5,"kind":65536,"name":"__type","url":"modules/_lib_fetch_.html#my_fetch.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"lib/fetch\".my_fetch"},{"id":6,"kind":64,"name":"fetch_resource","url":"modules/_lib_fetch_.html#fetch_resource","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":7,"kind":64,"name":"fetch_type","url":"modules/_lib_fetch_.html#fetch_type","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":8,"kind":64,"name":"fetch_html","url":"modules/_lib_fetch_.html#fetch_html","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":9,"kind":1,"name":"\"lib/opf\"","url":"modules/_lib_opf_.html","classes":"tsd-kind-external-module"},{"id":10,"kind":256,"name":"Package","url":"interfaces/_lib_opf_.package.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/opf\""},{"id":11,"kind":1024,"name":"package","url":"interfaces/_lib_opf_.package.html#package","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/opf\".Package"},{"id":12,"kind":128,"name":"PackageWrapper","url":"classes/_lib_opf_.packagewrapper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"lib/opf\""},{"id":13,"kind":512,"name":"constructor","url":"classes/_lib_opf_.packagewrapper.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":14,"kind":2048,"name":"add_manifest_item","url":"classes/_lib_opf_.packagewrapper.html#add_manifest_item","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":15,"kind":2048,"name":"add_creators","url":"classes/_lib_opf_.packagewrapper.html#add_creators","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":16,"kind":2048,"name":"add_dates","url":"classes/_lib_opf_.packagewrapper.html#add_dates","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":17,"kind":2048,"name":"serialize","url":"classes/_lib_opf_.packagewrapper.html#serialize","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":18,"kind":1,"name":"\"lib/css\"","url":"modules/_lib_css_.html","classes":"tsd-kind-external-module"},{"id":19,"kind":256,"name":"specStatus_css_mappings","url":"interfaces/_lib_css_.specstatus_css_mappings.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":20,"kind":1024,"name":"watermark","url":"interfaces/_lib_css_.specstatus_css_mappings.html#watermark","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":21,"kind":1024,"name":"logo_name","url":"interfaces/_lib_css_.specstatus_css_mappings.html#logo_name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":22,"kind":1024,"name":"logo_media_type","url":"interfaces/_lib_css_.specstatus_css_mappings.html#logo_media_type","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":23,"kind":1024,"name":"special_template","url":"interfaces/_lib_css_.specstatus_css_mappings.html#special_template","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":24,"kind":256,"name":"specStatus_mapping","url":"interfaces/_lib_css_.specstatus_mapping.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":25,"kind":32,"name":"extra_css","url":"modules/_lib_css_.html#extra_css","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":26,"kind":32,"name":"specStatus_simple","url":"modules/_lib_css_.html#specstatus_simple","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":27,"kind":2097152,"name":"specStatus_css","url":"modules/_lib_css_.html#specstatus_css","classes":"tsd-kind-object-literal tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":28,"kind":2097152,"name":"UNOFFICIAL","url":"modules/_lib_css_.html#specstatus_css.unofficial","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":29,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.unofficial.watermark-7","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":30,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.unofficial.logo_name-7","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":31,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.unofficial.logo_media_type-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":32,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.unofficial.special_template-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":33,"kind":2097152,"name":"FPWD","url":"modules/_lib_css_.html#specstatus_css.fpwd","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":34,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.fpwd.watermark-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD"},{"id":35,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.fpwd.logo_name-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD"},{"id":36,"kind":2097152,"name":"LC","url":"modules/_lib_css_.html#specstatus_css.lc","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":37,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.lc.watermark-6","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.LC"},{"id":38,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.lc.logo_name-6","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.LC"},{"id":39,"kind":2097152,"name":"FPWD-NOTE","url":"modules/_lib_css_.html#specstatus_css.fpwd_note","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":40,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.fpwd_note.watermark-5","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD-NOTE"},{"id":41,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.fpwd_note.logo_name-5","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD-NOTE"},{"id":42,"kind":2097152,"name":"WG-NOTE","url":"modules/_lib_css_.html#specstatus_css.wg_note","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":43,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.wg_note.watermark-8","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.WG-NOTE"},{"id":44,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.wg_note.logo_name-8","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.WG-NOTE"},{"id":45,"kind":2097152,"name":"BG-DRAFT","url":"modules/_lib_css_.html#specstatus_css.bg_draft","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":46,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.bg_draft.watermark","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":47,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.bg_draft.logo_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":48,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.bg_draft.logo_media_type","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":49,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.bg_draft.special_template","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":50,"kind":2097152,"name":"BG-FINAL","url":"modules/_lib_css_.html#specstatus_css.bg_final","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":51,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.bg_final.watermark-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":52,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.bg_final.logo_name-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":53,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.bg_final.logo_media_type-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":54,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.bg_final.special_template-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":55,"kind":2097152,"name":"CG-DRAFT","url":"modules/_lib_css_.html#specstatus_css.cg_draft","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":56,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.cg_draft.watermark-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":57,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.cg_draft.logo_name-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":58,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.cg_draft.logo_media_type-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":59,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.cg_draft.special_template-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":60,"kind":2097152,"name":"CG-FINAL","url":"modules/_lib_css_.html#specstatus_css.cg_final","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":61,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.cg_final.watermark-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":62,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.cg_final.logo_name-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":63,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.cg_final.logo_media_type-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":64,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.cg_final.special_template-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":65,"kind":64,"name":"extract_css","url":"modules/_lib_css_.html#extract_css","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/css\""},{"id":66,"kind":1,"name":"\"lib/xhtml\"","url":"modules/_lib_xhtml_.html","classes":"tsd-kind-external-module"},{"id":67,"kind":64,"name":"convert_dom","url":"modules/_lib_xhtml_.html#convert_dom","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/xhtml\""},{"id":68,"kind":64,"name":"convert_text","url":"modules/_lib_xhtml_.html#convert_text","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/xhtml\""},{"id":69,"kind":1,"name":"\"lib/nav\"","url":"modules/_lib_nav_.html","classes":"tsd-kind-external-module"},{"id":70,"kind":64,"name":"create_nav_file","url":"modules/_lib_nav_.html#create_nav_file","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/nav\""},{"id":71,"kind":1,"name":"\"lib/ocf\"","url":"modules/_lib_ocf_.html","classes":"tsd-kind-external-module"},{"id":72,"kind":128,"name":"OCF","url":"classes/_lib_ocf_.ocf.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"lib/ocf\""},{"id":73,"kind":1024,"name":"book","url":"classes/_lib_ocf_.ocf.html#book","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private","parent":"\"lib/ocf\".OCF"},{"id":74,"kind":512,"name":"constructor","url":"classes/_lib_ocf_.ocf.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":75,"kind":2048,"name":"append","url":"classes/_lib_ocf_.ocf.html#append","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":76,"kind":2048,"name":"finalize","url":"classes/_lib_ocf_.ocf.html#finalize","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":77,"kind":32,"name":"container_xml","url":"modules/_lib_ocf_.html#container_xml","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/ocf\""},{"id":78,"kind":1,"name":"\"lib/process\"","url":"modules/_lib_process_.html","classes":"tsd-kind-external-module"},{"id":79,"kind":256,"name":"ConfigOptions","url":"interfaces/_lib_process_.configoptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":80,"kind":256,"name":"DebugOptions","url":"interfaces/_lib_process_.debugoptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":81,"kind":1024,"name":"package","url":"interfaces/_lib_process_.debugoptions.html#package","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".DebugOptions"},{"id":82,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.debugoptions.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".DebugOptions"},{"id":83,"kind":256,"name":"Arguments","url":"interfaces/_lib_process_.arguments.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":84,"kind":1024,"name":"url","url":"interfaces/_lib_process_.arguments.html#url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":85,"kind":1024,"name":"respec","url":"interfaces/_lib_process_.arguments.html#respec","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":86,"kind":1024,"name":"config","url":"interfaces/_lib_process_.arguments.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":87,"kind":1024,"name":"package","url":"interfaces/_lib_process_.arguments.html#package","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"lib/process\".Arguments"},{"id":88,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.arguments.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"lib/process\".Arguments"},{"id":89,"kind":256,"name":"ResourceRef","url":"interfaces/_lib_process_.resourceref.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":90,"kind":1024,"name":"relative_url","url":"interfaces/_lib_process_.resourceref.html#relative_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":91,"kind":1024,"name":"media_type","url":"interfaces/_lib_process_.resourceref.html#media_type","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":92,"kind":1024,"name":"absolute_url","url":"interfaces/_lib_process_.resourceref.html#absolute_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":93,"kind":1024,"name":"text_content","url":"interfaces/_lib_process_.resourceref.html#text_content","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":94,"kind":1024,"name":"id","url":"interfaces/_lib_process_.resourceref.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":95,"kind":1024,"name":"properties","url":"interfaces/_lib_process_.resourceref.html#properties","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":96,"kind":256,"name":"Global","url":"interfaces/_lib_process_.global.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":97,"kind":1024,"name":"document_url","url":"interfaces/_lib_process_.global.html#document_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":98,"kind":1024,"name":"dom","url":"interfaces/_lib_process_.global.html#dom","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":99,"kind":1024,"name":"html_element","url":"interfaces/_lib_process_.global.html#html_element","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":100,"kind":1024,"name":"config","url":"interfaces/_lib_process_.global.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":101,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.global.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":102,"kind":1024,"name":"package","url":"interfaces/_lib_process_.global.html#package","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":103,"kind":1024,"name":"opf_content","url":"interfaces/_lib_process_.global.html#opf_content","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":104,"kind":1024,"name":"resources","url":"interfaces/_lib_process_.global.html#resources","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":105,"kind":256,"name":"LocalLinks","url":"interfaces/_lib_process_.locallinks.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":106,"kind":1024,"name":"query","url":"interfaces/_lib_process_.locallinks.html#query","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/process\".LocalLinks"},{"id":107,"kind":1024,"name":"attr","url":"interfaces/_lib_process_.locallinks.html#attr","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/process\".LocalLinks"},{"id":108,"kind":256,"name":"spec_generator","url":"interfaces/_lib_process_.spec_generator.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":109,"kind":1024,"name":"run","url":"interfaces/_lib_process_.spec_generator.html#run","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".spec_generator"},{"id":110,"kind":1024,"name":"config","url":"interfaces/_lib_process_.spec_generator.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".spec_generator"},{"id":111,"kind":32,"name":"spec_generator","url":"modules/_lib_process_.html#spec_generator-1","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":112,"kind":32,"name":"resource_references","url":"modules/_lib_process_.html#resource_references","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":113,"kind":64,"name":"create_epub","url":"modules/_lib_process_.html#create_epub","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":114,"kind":64,"name":"create_epub_from_dom","url":"modules/_lib_process_.html#create_epub_from_dom","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":115,"kind":64,"name":"get_extra_resources","url":"modules/_lib_process_.html#get_extra_resources","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":116,"kind":64,"name":"generate_overview_item","url":"modules/_lib_process_.html#generate_overview_item","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":117,"kind":64,"name":"generate_epub","url":"modules/_lib_process_.html#generate_epub","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":118,"kind":1,"name":"\"lib/cover\"","url":"modules/_lib_cover_.html","classes":"tsd-kind-external-module"},{"id":119,"kind":64,"name":"create_cover_page","url":"modules/_lib_cover_.html#create_cover_page","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/cover\""},{"id":120,"kind":1,"name":"\"main\"","url":"modules/_main_.html","classes":"tsd-kind-external-module"},{"id":121,"kind":64,"name":"main","url":"modules/_main_.html#main","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"main\""}]};
+            typedoc.search.data = {"kinds":{"1":"External module","32":"Variable","64":"Function","128":"Class","256":"Interface","512":"Constructor","1024":"Property","2048":"Method","65536":"Type literal","2097152":"Object literal","4194304":"Type alias"},"rows":[{"id":0,"kind":1,"name":"\"lib/fetch\"","url":"modules/_lib_fetch_.html","classes":"tsd-kind-external-module"},{"id":1,"kind":4194304,"name":"URL","url":"modules/_lib_fetch_.html#url","classes":"tsd-kind-type-alias tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":2,"kind":32,"name":"text_content","url":"modules/_lib_fetch_.html#text_content","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":3,"kind":64,"name":"check_Web_url","url":"modules/_lib_fetch_.html#check_web_url","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":4,"kind":32,"name":"my_fetch","url":"modules/_lib_fetch_.html#my_fetch","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/fetch\""},{"id":5,"kind":65536,"name":"__type","url":"modules/_lib_fetch_.html#my_fetch.__type","classes":"tsd-kind-type-literal tsd-parent-kind-variable tsd-is-not-exported","parent":"\"lib/fetch\".my_fetch"},{"id":6,"kind":64,"name":"fetch_resource","url":"modules/_lib_fetch_.html#fetch_resource","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":7,"kind":64,"name":"fetch_type","url":"modules/_lib_fetch_.html#fetch_type","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":8,"kind":64,"name":"fetch_html","url":"modules/_lib_fetch_.html#fetch_html","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/fetch\""},{"id":9,"kind":1,"name":"\"lib/opf\"","url":"modules/_lib_opf_.html","classes":"tsd-kind-external-module"},{"id":10,"kind":256,"name":"Package","url":"interfaces/_lib_opf_.package.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/opf\""},{"id":11,"kind":1024,"name":"package","url":"interfaces/_lib_opf_.package.html#package","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/opf\".Package"},{"id":12,"kind":128,"name":"PackageWrapper","url":"classes/_lib_opf_.packagewrapper.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"lib/opf\""},{"id":13,"kind":512,"name":"constructor","url":"classes/_lib_opf_.packagewrapper.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":14,"kind":2048,"name":"add_manifest_item","url":"classes/_lib_opf_.packagewrapper.html#add_manifest_item","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":15,"kind":2048,"name":"add_creators","url":"classes/_lib_opf_.packagewrapper.html#add_creators","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":16,"kind":2048,"name":"add_dates","url":"classes/_lib_opf_.packagewrapper.html#add_dates","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":17,"kind":2048,"name":"serialize","url":"classes/_lib_opf_.packagewrapper.html#serialize","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/opf\".PackageWrapper"},{"id":18,"kind":1,"name":"\"lib/css\"","url":"modules/_lib_css_.html","classes":"tsd-kind-external-module"},{"id":19,"kind":256,"name":"specStatus_css_mappings","url":"interfaces/_lib_css_.specstatus_css_mappings.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":20,"kind":1024,"name":"watermark","url":"interfaces/_lib_css_.specstatus_css_mappings.html#watermark","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":21,"kind":1024,"name":"logo_name","url":"interfaces/_lib_css_.specstatus_css_mappings.html#logo_name","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":22,"kind":1024,"name":"logo_media_type","url":"interfaces/_lib_css_.specstatus_css_mappings.html#logo_media_type","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":23,"kind":1024,"name":"special_template","url":"interfaces/_lib_css_.specstatus_css_mappings.html#special_template","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/css\".specStatus_css_mappings"},{"id":24,"kind":256,"name":"specStatus_mapping","url":"interfaces/_lib_css_.specstatus_mapping.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":25,"kind":32,"name":"specStatus_simple","url":"modules/_lib_css_.html#specstatus_simple","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":26,"kind":2097152,"name":"specStatus_css","url":"modules/_lib_css_.html#specstatus_css","classes":"tsd-kind-object-literal tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/css\""},{"id":27,"kind":2097152,"name":"UNOFFICIAL","url":"modules/_lib_css_.html#specstatus_css.unofficial","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":28,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.unofficial.watermark-7","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":29,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.unofficial.logo_name-7","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":30,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.unofficial.logo_media_type-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":31,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.unofficial.special_template-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.UNOFFICIAL"},{"id":32,"kind":2097152,"name":"FPWD","url":"modules/_lib_css_.html#specstatus_css.fpwd","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":33,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.fpwd.watermark-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD"},{"id":34,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.fpwd.logo_name-4","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD"},{"id":35,"kind":2097152,"name":"LC","url":"modules/_lib_css_.html#specstatus_css.lc","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":36,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.lc.watermark-6","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.LC"},{"id":37,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.lc.logo_name-6","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.LC"},{"id":38,"kind":2097152,"name":"FPWD-NOTE","url":"modules/_lib_css_.html#specstatus_css.fpwd_note","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":39,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.fpwd_note.watermark-5","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD-NOTE"},{"id":40,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.fpwd_note.logo_name-5","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.FPWD-NOTE"},{"id":41,"kind":2097152,"name":"WG-NOTE","url":"modules/_lib_css_.html#specstatus_css.wg_note","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":42,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.wg_note.watermark-8","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.WG-NOTE"},{"id":43,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.wg_note.logo_name-8","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.WG-NOTE"},{"id":44,"kind":2097152,"name":"BG-DRAFT","url":"modules/_lib_css_.html#specstatus_css.bg_draft","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":45,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.bg_draft.watermark","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":46,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.bg_draft.logo_name","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":47,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.bg_draft.logo_media_type","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":48,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.bg_draft.special_template","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-DRAFT"},{"id":49,"kind":2097152,"name":"BG-FINAL","url":"modules/_lib_css_.html#specstatus_css.bg_final","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":50,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.bg_final.watermark-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":51,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.bg_final.logo_name-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":52,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.bg_final.logo_media_type-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":53,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.bg_final.special_template-1","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.BG-FINAL"},{"id":54,"kind":2097152,"name":"CG-DRAFT","url":"modules/_lib_css_.html#specstatus_css.cg_draft","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":55,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.cg_draft.watermark-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":56,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.cg_draft.logo_name-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":57,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.cg_draft.logo_media_type-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":58,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.cg_draft.special_template-2","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-DRAFT"},{"id":59,"kind":2097152,"name":"CG-FINAL","url":"modules/_lib_css_.html#specstatus_css.cg_final","classes":"tsd-kind-object-literal tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css"},{"id":60,"kind":32,"name":"watermark","url":"modules/_lib_css_.html#specstatus_css.cg_final.watermark-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":61,"kind":32,"name":"logo_name","url":"modules/_lib_css_.html#specstatus_css.cg_final.logo_name-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":62,"kind":32,"name":"logo_media_type","url":"modules/_lib_css_.html#specstatus_css.cg_final.logo_media_type-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":63,"kind":32,"name":"special_template","url":"modules/_lib_css_.html#specstatus_css.cg_final.special_template-3","classes":"tsd-kind-variable tsd-parent-kind-object-literal tsd-is-not-exported","parent":"\"lib/css\".specStatus_css.CG-FINAL"},{"id":64,"kind":64,"name":"extract_css","url":"modules/_lib_css_.html#extract_css","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/css\""},{"id":65,"kind":1,"name":"\"lib/xhtml\"","url":"modules/_lib_xhtml_.html","classes":"tsd-kind-external-module"},{"id":66,"kind":64,"name":"convert_dom","url":"modules/_lib_xhtml_.html#convert_dom","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/xhtml\""},{"id":67,"kind":64,"name":"convert_text","url":"modules/_lib_xhtml_.html#convert_text","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/xhtml\""},{"id":68,"kind":1,"name":"\"lib/nav\"","url":"modules/_lib_nav_.html","classes":"tsd-kind-external-module"},{"id":69,"kind":64,"name":"create_nav_file","url":"modules/_lib_nav_.html#create_nav_file","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/nav\""},{"id":70,"kind":1,"name":"\"lib/ocf\"","url":"modules/_lib_ocf_.html","classes":"tsd-kind-external-module"},{"id":71,"kind":128,"name":"OCF","url":"classes/_lib_ocf_.ocf.html","classes":"tsd-kind-class tsd-parent-kind-external-module","parent":"\"lib/ocf\""},{"id":72,"kind":1024,"name":"book","url":"classes/_lib_ocf_.ocf.html#book","classes":"tsd-kind-property tsd-parent-kind-class tsd-is-private","parent":"\"lib/ocf\".OCF"},{"id":73,"kind":512,"name":"constructor","url":"classes/_lib_ocf_.ocf.html#constructor","classes":"tsd-kind-constructor tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":74,"kind":2048,"name":"append","url":"classes/_lib_ocf_.ocf.html#append","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":75,"kind":2048,"name":"finalize","url":"classes/_lib_ocf_.ocf.html#finalize","classes":"tsd-kind-method tsd-parent-kind-class","parent":"\"lib/ocf\".OCF"},{"id":76,"kind":32,"name":"container_xml","url":"modules/_lib_ocf_.html#container_xml","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/ocf\""},{"id":77,"kind":1,"name":"\"lib/overview\"","url":"modules/_lib_overview_.html","classes":"tsd-kind-external-module"},{"id":78,"kind":64,"name":"generate_overview_item","url":"modules/_lib_overview_.html#generate_overview_item","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/overview\""},{"id":79,"kind":1,"name":"\"lib/process\"","url":"modules/_lib_process_.html","classes":"tsd-kind-external-module"},{"id":80,"kind":256,"name":"ConfigOptions","url":"interfaces/_lib_process_.configoptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":81,"kind":256,"name":"DebugOptions","url":"interfaces/_lib_process_.debugoptions.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":82,"kind":1024,"name":"package","url":"interfaces/_lib_process_.debugoptions.html#package","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".DebugOptions"},{"id":83,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.debugoptions.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".DebugOptions"},{"id":84,"kind":256,"name":"Arguments","url":"interfaces/_lib_process_.arguments.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":85,"kind":1024,"name":"url","url":"interfaces/_lib_process_.arguments.html#url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":86,"kind":1024,"name":"respec","url":"interfaces/_lib_process_.arguments.html#respec","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":87,"kind":1024,"name":"config","url":"interfaces/_lib_process_.arguments.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Arguments"},{"id":88,"kind":1024,"name":"package","url":"interfaces/_lib_process_.arguments.html#package","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"lib/process\".Arguments"},{"id":89,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.arguments.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-inherited","parent":"\"lib/process\".Arguments"},{"id":90,"kind":256,"name":"ResourceRef","url":"interfaces/_lib_process_.resourceref.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":91,"kind":1024,"name":"relative_url","url":"interfaces/_lib_process_.resourceref.html#relative_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":92,"kind":1024,"name":"media_type","url":"interfaces/_lib_process_.resourceref.html#media_type","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":93,"kind":1024,"name":"absolute_url","url":"interfaces/_lib_process_.resourceref.html#absolute_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":94,"kind":1024,"name":"text_content","url":"interfaces/_lib_process_.resourceref.html#text_content","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":95,"kind":1024,"name":"id","url":"interfaces/_lib_process_.resourceref.html#id","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":96,"kind":1024,"name":"properties","url":"interfaces/_lib_process_.resourceref.html#properties","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".ResourceRef"},{"id":97,"kind":256,"name":"Global","url":"interfaces/_lib_process_.global.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":98,"kind":1024,"name":"document_url","url":"interfaces/_lib_process_.global.html#document_url","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":99,"kind":1024,"name":"dom","url":"interfaces/_lib_process_.global.html#dom","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":100,"kind":1024,"name":"html_element","url":"interfaces/_lib_process_.global.html#html_element","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":101,"kind":1024,"name":"config","url":"interfaces/_lib_process_.global.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":102,"kind":1024,"name":"trace","url":"interfaces/_lib_process_.global.html#trace","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":103,"kind":1024,"name":"package","url":"interfaces/_lib_process_.global.html#package","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":104,"kind":1024,"name":"opf_content","url":"interfaces/_lib_process_.global.html#opf_content","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":105,"kind":1024,"name":"resources","url":"interfaces/_lib_process_.global.html#resources","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".Global"},{"id":106,"kind":256,"name":"LocalLinks","url":"interfaces/_lib_process_.locallinks.html","classes":"tsd-kind-interface tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":107,"kind":1024,"name":"query","url":"interfaces/_lib_process_.locallinks.html#query","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/process\".LocalLinks"},{"id":108,"kind":1024,"name":"attr","url":"interfaces/_lib_process_.locallinks.html#attr","classes":"tsd-kind-property tsd-parent-kind-interface tsd-is-not-exported","parent":"\"lib/process\".LocalLinks"},{"id":109,"kind":256,"name":"spec_generator","url":"interfaces/_lib_process_.spec_generator.html","classes":"tsd-kind-interface tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":110,"kind":1024,"name":"run","url":"interfaces/_lib_process_.spec_generator.html#run","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".spec_generator"},{"id":111,"kind":1024,"name":"config","url":"interfaces/_lib_process_.spec_generator.html#config","classes":"tsd-kind-property tsd-parent-kind-interface","parent":"\"lib/process\".spec_generator"},{"id":112,"kind":32,"name":"spec_generator","url":"modules/_lib_process_.html#spec_generator-1","classes":"tsd-kind-variable tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":113,"kind":32,"name":"resource_references","url":"modules/_lib_process_.html#resource_references","classes":"tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":114,"kind":64,"name":"create_epub","url":"modules/_lib_process_.html#create_epub","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":115,"kind":64,"name":"create_epub_from_dom","url":"modules/_lib_process_.html#create_epub_from_dom","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/process\""},{"id":116,"kind":64,"name":"get_extra_resources","url":"modules/_lib_process_.html#get_extra_resources","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":117,"kind":64,"name":"generate_epub","url":"modules/_lib_process_.html#generate_epub","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"lib/process\""},{"id":118,"kind":1,"name":"\"lib/cover\"","url":"modules/_lib_cover_.html","classes":"tsd-kind-external-module"},{"id":119,"kind":64,"name":"create_cover_page","url":"modules/_lib_cover_.html#create_cover_page","classes":"tsd-kind-function tsd-parent-kind-external-module","parent":"\"lib/cover\""},{"id":120,"kind":1,"name":"\"main\"","url":"modules/_main_.html","classes":"tsd-kind-external-module"},{"id":121,"kind":64,"name":"main","url":"modules/_main_.html#main","classes":"tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported","parent":"\"main\""}]};
