+++ docs/classes/_lib_opf_.packagewrapper.html
@@ -119,7 +119,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/opf.ts#L165">lib/opf.ts:165</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/opf.ts#L165">lib/opf.ts:165</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -160,7 +160,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/opf.ts#L248">lib/opf.ts:248</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/opf.ts#L248">lib/opf.ts:248</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -191,7 +191,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/opf.ts#L269">lib/opf.ts:269</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/opf.ts#L269">lib/opf.ts:269</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -222,7 +222,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/opf.ts#L236">lib/opf.ts:236</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/opf.ts#L236">lib/opf.ts:236</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -253,7 +253,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/opf.ts#L285">lib/opf.ts:285</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/opf.ts#L285">lib/opf.ts:285</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -293,6 +293,9 @@
 					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="../modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_lib_process_.html">&quot;lib/process&quot;</a>
