+++ docs/index.html
@@ -112,6 +112,9 @@
 						<a href="modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
+						<a href="modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
 						<a href="modules/_lib_process_.html">&quot;lib/process&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
