+++ docs/interfaces/_lib_css_.specstatus_mapping.html
@@ -119,6 +119,9 @@
 						<a href="../modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
+						<a href="../modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
 						<a href="../modules/_lib_process_.html">&quot;lib/process&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
@@ -141,9 +144,6 @@
 					</li>
 				</ul>
 				<ul class="after-current">
-					<li class=" tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported">
-						<a href="../modules/_lib_css_.html#extra_css" class="tsd-kind-icon">extra_<wbr>css</a>
-					</li>
 					<li class=" tsd-kind-variable tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_css_.html#specstatus_simple" class="tsd-kind-icon">spec<wbr>Status_<wbr>simple</a>
 					</li>
