+++ docs/interfaces/_lib_process_.debugoptions.html
@@ -111,7 +111,7 @@
 					<div class="tsd-signature tsd-kind-icon">package<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L53">lib/process.ts:53</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L53">lib/process.ts:53</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -126,7 +126,7 @@
 					<div class="tsd-signature tsd-kind-icon">trace<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L58">lib/process.ts:58</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L58">lib/process.ts:58</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -160,6 +160,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="../modules/_lib_process_.html">&quot;lib/process&quot;</a>
@@ -221,9 +224,6 @@
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#generate_epub" class="tsd-kind-icon">generate_<wbr>epub</a>
-					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
-						<a href="../modules/_lib_process_.html#generate_overview_item" class="tsd-kind-icon">generate_<wbr>overview_<wbr>item</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#get_extra_resources" class="tsd-kind-icon">get_<wbr>extra_<wbr>resources</a>
