+++ docs/interfaces/_lib_process_.global.html
@@ -112,7 +112,7 @@
 					<div class="tsd-signature tsd-kind-icon">config<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">any</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L127">lib/process.ts:127</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L127">lib/process.ts:127</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -128,7 +128,7 @@
 					<div class="tsd-signature tsd-kind-icon">document_<wbr>url<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L115">lib/process.ts:115</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L115">lib/process.ts:115</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -143,7 +143,7 @@
 					<div class="tsd-signature tsd-kind-icon">dom<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">JSDOM</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L118">lib/process.ts:118</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L118">lib/process.ts:118</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -158,7 +158,7 @@
 					<div class="tsd-signature tsd-kind-icon">html_<wbr>element<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">Element</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L121">lib/process.ts:121</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L121">lib/process.ts:121</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -173,7 +173,7 @@
 					<div class="tsd-signature tsd-kind-icon">opf_<wbr>content<span class="tsd-signature-symbol">:</span> <a href="../classes/_lib_opf_.packagewrapper.html" class="tsd-signature-type">PackageWrapper</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L142">lib/process.ts:142</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L142">lib/process.ts:142</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -188,7 +188,7 @@
 					<div class="tsd-signature tsd-kind-icon">package<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L137">lib/process.ts:137</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L137">lib/process.ts:137</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -203,7 +203,7 @@
 					<div class="tsd-signature tsd-kind-icon">resources<span class="tsd-signature-symbol">:</span> <a href="_lib_process_.resourceref.html" class="tsd-signature-type">ResourceRef</a><span class="tsd-signature-symbol">[]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L150">lib/process.ts:150</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L150">lib/process.ts:150</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -221,7 +221,7 @@
 					<div class="tsd-signature tsd-kind-icon">trace<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">boolean</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L132">lib/process.ts:132</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L132">lib/process.ts:132</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -255,6 +255,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="../modules/_lib_process_.html">&quot;lib/process&quot;</a>
@@ -334,9 +337,6 @@
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#generate_epub" class="tsd-kind-icon">generate_<wbr>epub</a>
-					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
-						<a href="../modules/_lib_process_.html#generate_overview_item" class="tsd-kind-icon">generate_<wbr>overview_<wbr>item</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#get_extra_resources" class="tsd-kind-icon">get_<wbr>extra_<wbr>resources</a>
