+++ docs/interfaces/_lib_process_.resourceref.html
@@ -110,7 +110,7 @@
 					<div class="tsd-signature tsd-kind-icon">absolute_<wbr>url<span class="tsd-signature-symbol">:</span> <a href="../modules/_lib_fetch_.html#url" class="tsd-signature-type">URL</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L99">lib/process.ts:99</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L99">lib/process.ts:99</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -125,7 +125,7 @@
 					<div class="tsd-signature tsd-kind-icon">id<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L105">lib/process.ts:105</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L105">lib/process.ts:105</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -140,7 +140,7 @@
 					<div class="tsd-signature tsd-kind-icon">media_<wbr>type<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L96">lib/process.ts:96</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L96">lib/process.ts:96</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -155,7 +155,7 @@
 					<div class="tsd-signature tsd-kind-icon">properties<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L108">lib/process.ts:108</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L108">lib/process.ts:108</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -170,7 +170,7 @@
 					<div class="tsd-signature tsd-kind-icon">relative_<wbr>url<span class="tsd-signature-symbol">:</span> <a href="../modules/_lib_fetch_.html#url" class="tsd-signature-type">URL</a></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L93">lib/process.ts:93</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L93">lib/process.ts:93</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -185,7 +185,7 @@
 					<div class="tsd-signature tsd-kind-icon">text_<wbr>content<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/process.ts#L102">lib/process.ts:102</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/process.ts#L102">lib/process.ts:102</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -219,6 +219,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="../modules/_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="../modules/_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class="current tsd-kind-external-module">
 						<a href="../modules/_lib_process_.html">&quot;lib/process&quot;</a>
@@ -292,9 +295,6 @@
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#generate_epub" class="tsd-kind-icon">generate_<wbr>epub</a>
-					</li>
-					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
-						<a href="../modules/_lib_process_.html#generate_overview_item" class="tsd-kind-icon">generate_<wbr>overview_<wbr>item</a>
 					</li>
 					<li class=" tsd-kind-function tsd-parent-kind-external-module tsd-is-not-exported">
 						<a href="../modules/_lib_process_.html#get_extra_resources" class="tsd-kind-icon">get_<wbr>extra_<wbr>resources</a>
