+++ docs/modules/_lib_cover_.html
@@ -101,7 +101,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/cover.ts#L83">lib/cover.ts:83</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/cover.ts#L83">lib/cover.ts:83</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -147,6 +147,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_process_.html">&quot;lib/process&quot;</a>
