+++ docs/modules/_lib_fetch_.html
@@ -113,7 +113,7 @@
 					<div class="tsd-signature tsd-kind-icon">URL<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L8">lib/fetch.ts:8</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L8">lib/fetch.ts:8</a></li>
 						</ul>
 					</aside>
 				</section>
@@ -126,7 +126,7 @@
 					<div class="tsd-signature tsd-kind-icon">my_<wbr>fetch<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-symbol">(</span>arg<span class="tsd-signature-symbol">: </span><span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">)</span><span class="tsd-signature-symbol"> =&gt; </span><span class="tsd-signature-type">Promise</span><span class="tsd-signature-symbol">&lt;</span><span class="tsd-signature-type">any</span><span class="tsd-signature-symbol">&gt;</span><span class="tsd-signature-symbol"> = (process !&#x3D;&#x3D; undefined) ? node_fetch.default : fetch</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L136">lib/fetch.ts:136</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L136">lib/fetch.ts:136</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -167,7 +167,7 @@
 					<div class="tsd-signature tsd-kind-icon">text_<wbr>content<span class="tsd-signature-symbol">:</span> <span class="tsd-signature-type">string</span><span class="tsd-signature-symbol">[]</span><span class="tsd-signature-symbol"> = [json_media_type,jsonld_media_type,html_media_type,xhtml_media_type,css_media_type,svg_media_type,js_media_type,es_media_type]</span></div>
 					<aside class="tsd-sources">
 						<ul>
-							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L61">lib/fetch.ts:61</a></li>
+							<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L61">lib/fetch.ts:61</a></li>
 						</ul>
 					</aside>
 					<div class="tsd-comment tsd-typography">
@@ -189,7 +189,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L91">lib/fetch.ts:91</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L91">lib/fetch.ts:91</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -234,7 +234,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L231">lib/fetch.ts:231</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L231">lib/fetch.ts:231</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -272,7 +272,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L147">lib/fetch.ts:147</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L147">lib/fetch.ts:147</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -313,7 +313,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/fetch.ts#L197">lib/fetch.ts:197</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/fetch.ts#L197">lib/fetch.ts:197</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -363,6 +363,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_process_.html">&quot;lib/process&quot;</a>
