+++ docs/modules/_lib_xhtml_.html
@@ -105,7 +105,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/xhtml.ts#L24">lib/xhtml.ts:24</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/xhtml.ts#L24">lib/xhtml.ts:24</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -139,7 +139,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/lib/xhtml.ts#L34">lib/xhtml.ts:34</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/lib/xhtml.ts#L34">lib/xhtml.ts:34</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -185,6 +185,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_process_.html">&quot;lib/process&quot;</a>
