+++ docs/modules/_main_.html
@@ -98,7 +98,7 @@
 						<li class="tsd-description">
 							<aside class="tsd-sources">
 								<ul>
-									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/dd7cccb/src/main.ts#L45">main.ts:45</a></li>
+									<li>Defined in <a href="https://github.com/iherman/respec-to-epub/blob/99b67b0/src/main.ts#L46">main.ts:46</a></li>
 								</ul>
 							</aside>
 							<div class="tsd-comment tsd-typography">
@@ -158,6 +158,9 @@
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_opf_.html">&quot;lib/opf&quot;</a>
+					</li>
+					<li class=" tsd-kind-external-module">
+						<a href="_lib_overview_.html">&quot;lib/overview&quot;</a>
 					</li>
 					<li class=" tsd-kind-external-module">
 						<a href="_lib_process_.html">&quot;lib/process&quot;</a>
